Cleanup quantization related changes

+++ hunyuan_model/models.py
@@ -634,7 +634,7 @@ class HYVideoDiffusionTransformer(nn.Module):  # ModelMixin, ConfigMixin):
         logger.info(f"Scaling transformer to FP8{' and enabling fp8_fast/mm_scaled' if use_mmscaled else ''}...")
         state_dict = self.state_dict()
         state_dict = optimize_state_dict_with_fp8(state_dict, device, target_layer_keys=["single_blocks", "double_blocks"], exclude_layer_keys=params_to_keep, move_to_device=move_to_device)
-        apply_fp8_monkey_patch(self, state_dict, use_scaled_mm=use_mmscaled, original_weight_dtype=self.dtype, upcast_linear=upcast_linear)
+        apply_fp8_monkey_patch(self, state_dict, use_scaled_mm=use_mmscaled, upcast_linear=upcast_linear)
         if move_to_device:
             logger.info(f"Moving weights to {load_device}")
             for key in state_dict.keys():
